+++ apps/meteor/client/meteor/login/cas.ts
@@ -1,7 +1,7 @@
 import { Random } from '@rocket.chat/random';
 import { Meteor } from 'meteor/meteor';
 
-import { callLoginMethod } from '../../../lib/2fa/overrideLoginMethod';
+import { callLoginMethod } from '../../lib/2fa/overrideLoginMethod';
 
 declare module 'meteor/meteor' {
 	// eslint-disable-next-line @typescript-eslint/no-namespace
@@ -12,7 +12,7 @@ declare module 'meteor/meteor' {
 
 Meteor.loginWithCas = (_, callback) => {
 	const credentialToken = Random.id();
-	import('../../../lib/openCASLoginPopup')
+	import('../../lib/openCASLoginPopup')
 		.then(({ openCASLoginPopup }) => openCASLoginPopup(credentialToken))
 		.then(() => callLoginMethod({ methodArguments: [{ cas: { credentialToken } }] }))
 		.then(() => callback?.())
