+++ apps/meteor/client/meteor/login/google.ts
@@ -7,8 +7,8 @@ import { Meteor } from 'meteor/meteor';
 import { OAuth } from 'meteor/oauth';
 
 import { createOAuthTotpLoginMethod } from './oauth';
-import { overrideLoginMethod, type LoginCallback } from '../../../lib/2fa/overrideLoginMethod';
-import { wrapRequestCredentialFn } from '../../../lib/wrapRequestCredentialFn';
+import { overrideLoginMethod, type LoginCallback } from '../../lib/2fa/overrideLoginMethod';
+import { wrapRequestCredentialFn } from '../../lib/wrapRequestCredentialFn';
 
 declare module 'meteor/meteor' {
 	// eslint-disable-next-line @typescript-eslint/no-namespace
