+++ apps/meteor/client/meteor/login/index.ts
@@ -1,0 +1,11 @@
+import './cas';
+import './crowd';
+import './facebook';
+import './github';
+import './google';
+import './ldap';
+import './meteorDeveloperAccount';
+import './oauth';
+import './password';
+import './saml';
+import './twitter';
