+++ apps/meteor/client/meteor/login/meteorDeveloperAccount.ts
@@ -5,8 +5,8 @@ import { MeteorDeveloperAccounts } from 'meteor/meteor-developer-oauth';
 import { OAuth } from 'meteor/oauth';
 
 import { createOAuthTotpLoginMethod } from './oauth';
-import { overrideLoginMethod } from '../../../lib/2fa/overrideLoginMethod';
-import { wrapRequestCredentialFn } from '../../../lib/wrapRequestCredentialFn';
+import { overrideLoginMethod } from '../../lib/2fa/overrideLoginMethod';
+import { wrapRequestCredentialFn } from '../../lib/wrapRequestCredentialFn';
 
 const { loginWithMeteorDeveloperAccount } = Meteor;
 const loginWithMeteorDeveloperAccountAndTOTP = createOAuthTotpLoginMethod(MeteorDeveloperAccounts);
