+++ apps/meteor/client/meteor/login/saml.ts
@@ -2,8 +2,8 @@ import { Random } from '@rocket.chat/random';
 import { Accounts } from 'meteor/accounts-base';
 import { Meteor } from 'meteor/meteor';
 
-import { type LoginCallback, callLoginMethod, handleLogin } from '../../../lib/2fa/overrideLoginMethod';
-import { settings } from '../../../lib/settings';
+import { type LoginCallback, callLoginMethod, handleLogin } from '../../lib/2fa/overrideLoginMethod';
+import { settings } from '../../lib/settings';
 
 declare module 'meteor/meteor' {
 	// eslint-disable-next-line @typescript-eslint/no-namespace
@@ -61,7 +61,7 @@ Meteor.logout = async function (...args) {
 	if (provider && settings.peek('SAML_Custom_Default_idp_slo_redirect_url')) {
 		console.info('SAML session terminated via SLO');
 
-		const { sdk } = await import('../../../../app/utils/client/lib/SDKClient');
+		const { sdk } = await import('../../../app/utils/client/lib/SDKClient');
 		sdk
 			.call('samlLogout', provider)
 			.then((result) => {
