+++ apps/meteor/client/meteor/overrides/index.ts
@@ -1,18 +1,7 @@
-// import './ddpOverREST';
-// import './totpOnCall';
-// import './oauthRedirectUri';
-// import './unstoreLoginToken';
-// import './userAndUsers';
-import './login/cas';
-import './login/crowd';
-import './login/facebook';
-import './login/github';
-import './login/google';
-import './login/ldap';
-import './login/meteorDeveloperAccount';
-import './login/oauth';
-import './login/password';
-import './login/saml';
-import './login/twitter';
-// import './settings';
-// import './desktopInjection';
+import './ddpOverREST';
+import './desktopInjection';
+import './oauthRedirectUri';
+import './settings';
+import './totpOnCall';
+import './unstoreLoginToken';
+import './userAndUsers';
